requirements.txt
strands-agents==1.6.0
rich==14.1.0